Modify eviction struct list for inorder eviction

+++ bpf/types.h
@@ -160,7 +160,7 @@ typedef struct pkt_info_t {
 
 // Structure for payload metadata
 typedef struct payload_meta_t {
-    u16 if_index;
+    u32 if_index;
     u16 pkt_len;
     u64 timestamp;  // timestamp when packet received by ebpf
 } __attribute__((packed)) payload_meta;
